refactor(gptme-wisdom-mcp): rename package per maintainer feedback
