+++ packages/gptme-wisdom-mcp/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "gptme-rag-mcp"
+name = "gptme-wisdom-mcp"
 version = "0.1.0"
 description = "RAG-as-MCP Knowledge Server for gptme agents — search CS books and session history via MCP"
 authors = [
@@ -40,14 +40,14 @@ test = [
 ]
 
 [project.scripts]
-rag-mcp-server = "gptme_rag_mcp.mcp_server:main"
+gptme-wisdom-mcp = "gptme_wisdom_mcp.mcp_server:main"
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/gptme_rag_mcp"]
+packages = ["src/gptme_wisdom_mcp"]
 
 [tool.ruff]
 line-length = 100
