+++ packages/gptme-wisdom-mcp/src/gptme_wisdom_mcp/__init__.py
@@ -1,0 +1,3 @@
+"""gptme-wisdom-mcp: RAG-as-MCP Knowledge Server for gptme agents."""
+
+__version__ = "0.1.0"
