+++ packages/gptme-wisdom-mcp/src/gptme_wisdom_mcp/indexer.py
