+++ packages/gptme-wisdom-mcp/tests/test_indexer.py
@@ -1,4 +1,4 @@
-"""Basic smoke tests for gptme-rag-mcp indexer and parsers."""
+"""Basic smoke tests for gptme-wisdom-mcp indexer and parsers."""
 
 import json
 import tempfile
@@ -8,8 +8,8 @@ from pathlib import Path
 
 def test_session_index_roundtrip():
     """Create an index, add a document, search, find it."""
-    from gptme_rag_mcp.indexer import SessionIndex
-    from gptme_rag_mcp.parsers import SessionDocument
+    from gptme_wisdom_mcp.indexer import SessionIndex
+    from gptme_wisdom_mcp.parsers import SessionDocument
 
     with tempfile.TemporaryDirectory() as tmpdir:
         db = Path(tmpdir) / "test.db"
@@ -38,8 +38,8 @@ def test_session_index_roundtrip():
 
 def test_purge_rebuild_preserves_content_search():
     """Purging one path should not degrade FTS recall for remaining content."""
-    from gptme_rag_mcp.indexer import SessionIndex
-    from gptme_rag_mcp.parsers import SessionDocument
+    from gptme_wisdom_mcp.indexer import SessionIndex
+    from gptme_wisdom_mcp.parsers import SessionDocument
 
     with tempfile.TemporaryDirectory() as tmpdir:
         db = Path(tmpdir) / "test.db"
@@ -71,8 +71,8 @@ def test_purge_rebuild_preserves_content_search():
 
 def test_book_index_roundtrip():
     """Create a book index, add a chunk, search, find it."""
-    from gptme_rag_mcp.indexer import BookIndex
-    from gptme_rag_mcp.parsers import BookDocument
+    from gptme_wisdom_mcp.indexer import BookIndex
+    from gptme_wisdom_mcp.parsers import BookDocument
 
     with tempfile.TemporaryDirectory() as tmpdir:
         db = Path(tmpdir) / "wisdom.db"
@@ -99,7 +99,7 @@ def test_book_index_roundtrip():
 
 def test_parse_book_text_basic():
     """parse_book_text should chunk content and return BookDocuments."""
-    from gptme_rag_mcp.parsers import parse_book_text
+    from gptme_wisdom_mcp.parsers import parse_book_text
 
     text = """# Chapter 1: Introduction
 
@@ -127,8 +127,8 @@ Every expression evaluates to a value in the language.
 
 def test_mcp_server_importable():
     """mcp_server module should import without error."""
-    import gptme_rag_mcp.mcp_server  # noqa: F401
-    from gptme_rag_mcp.mcp_server import search_wisdom, search_sessions, list_wisdom_sources
+    import gptme_wisdom_mcp.mcp_server  # noqa: F401
+    from gptme_wisdom_mcp.mcp_server import search_wisdom, search_sessions, list_wisdom_sources
 
     assert callable(search_wisdom)
     assert callable(search_sessions)
